+++ internal/server/grow.go
@@ -175,6 +175,12 @@ func (h *hub) randomTickBlock(players map[int32]*tracked, dim, x, y, z int) {
 	if h.tickDripstone(players, dim, x, y, z, state) {
 		return
 	}
+	if h.tickEyeblossom(players, dim, x, y, z, state) {
+		return
+	}
+	if h.tickNetherPortal(players, dim, x, y, z, state) {
+		return
+	}
 	if h.tickRedstoneOre(players, dim, x, y, z, state) {
 		return
 	}